--- plugin/pymobiledevicelite/.bazelrc
@@ -1,1 +1,0 @@
-common --enable_bzlmod
