--- plugin/pymobiledevicelite/MODULE.bazel
@@ -1,23 +1,0 @@
-"""pymobiledeveice3 binary"""
-module(
-    name = "pymobiledevice3",
-    version = "0.0.1",
-)
-
-bazel_dep(name = "rules_python", version = "0.25.0")
-
-python = use_extension("@rules_python//python/extensions:python.bzl", "python")
-
-python.toolchain(
-    configure_coverage_tool = True,
-    python_version = "3.10",
-    is_default = True,
-)
-
-pip = use_extension("@rules_python//python/extensions:pip.bzl", "pip")
-pip.parse(
-    hub_name = "my_deps",
-    python_version = "3.10",
-    requirements_lock = "//:requirements_lock.txt",
-)
-use_repo(pip, "my_deps")
