--- plugin/pymobiledevicelite/requirements.in
@@ -1,3 +1,0 @@
---extra-index-url https://pypi.python.org/simple/
-
-pymobiledevice3==2.17.2
