+++ plugin/zxz-moe-bis/pymobiledevicelite/.bazelrc
@@ -1,0 +1,1 @@
+common --enable_bzlmod
