+++ plugin/zxz-moe-bis/pymobiledevicelite/BUILD.bazel
@@ -1,0 +1,26 @@
+"""
+This python program is a very simplified version of pymobiledevice3(https://github.com/doronz88/pymobiledevice3)
+"""
+
+load("@my_deps//:requirements.bzl", "requirement")
+load("@rules_python//python:defs.bzl", "py_binary")
+load("@rules_python//python:pip.bzl", "compile_pip_requirements")
+
+compile_pip_requirements(
+    name = "requirements",
+    extra_args = [
+        "--allow-unsafe",
+        "--resolver=backtracking",
+    ],
+    requirements_in = "requirements.in",
+    requirements_txt = "requirements_lock.txt",
+)
+
+py_binary(
+    name = "pymobiledevicelite",
+    srcs = ["pymobiledevicelite.py"],
+    visibility = ["//visibility:public"],
+    deps = [
+        requirement("pymobiledevice3"),
+    ],
+)
