+++ plugin/zxz-moe-bis/pymobiledevicelite/requirements.in
@@ -1,0 +1,3 @@
+--extra-index-url https://pypi.python.org/simple/
+
+pymobiledevice3==2.17.2
